Extracted get_song into rhyme_command
